Create Food_simple_listing
The query will group the product industries by name so the report results will fall under each industry title. ORDER BY and DESC will tell the query to order the output by count_reports in descending order so, the industry with the most amount of reports will appear at the top of the output table. LIMIT will limit the results to ten industries with the highest numbers of reports.