Create Food_listing_hospitalizations
The outer query used here is just like the previous query. SELECT calls the product industry names, and COUNT counts the number of reports and labels them as count_hospitalizations.  The query is told to pull this data from the food_events table. 

The subquery begins with a WHERE clause and an IN operator—these will allow the subquery to specify multiple values. The second SELECT statement tells the subquery to pull the product industry names FROM the food_events table in the FDA_food dataset. Results are then grouped by product industry name and ordered in descending order. LIMIT tells the query to limit the number of results to 10. AND and LIKE are used to indicate if hospitalizations are present within the outcomes field. 

The outer query is then closed by grouping results by product_industry_name and ordered in descending order by the number of hospitalizations recorded.